move wire package to utils
