+++ pkg/gds/store/leveldb/iterator.go
@@ -4,7 +4,7 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/syndtr/goleveldb/leveldb/iterator"
 	"github.com/trisacrypto/directory/pkg/gds/models/v1"
-	"github.com/trisacrypto/directory/pkg/gds/store/wire"
+	"github.com/trisacrypto/directory/pkg/utils/wire"
 	pb "github.com/trisacrypto/trisa/pkg/trisa/gds/models/v1beta1"
 	"google.golang.org/protobuf/proto"
 )
