+++ pkg/utils/wire/testdata/index::categories.json
