+++ pkg/utils/wire/testdata/index::names.json
