+++ pkg/utils/wire/testdata/peers::8.json
@@ -1,0 +1,9 @@
+{
+    "id": 8,
+    "addr": "localhost:4435",
+    "name": "frontier",
+    "region": "localhost",
+    "created": "2021-07-07T18:57:21-04:00",
+    "modified": "2021-07-07T18:57:21-04:00",
+    "extra": {}
+}
